requirements.txt
flask
flask_sqlalchemy 
flask_login 
flask_bcrypt 
flask_wtf 
wtforms 
email_validator
markdown
libsass